Intermediary step while debugging
